+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased]
 ### Added
 - Display test coverage in IDE
+### Changed
+- Auto-import packages when extracting a widget to a new file
 ### Deprecated
 - Deprecated bloc feature. Please use the official bloc plugin (https://plugins.jetbrains.com/plugin/12129-bloc)
 ### Fixed